scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.5.0
pandas==2.2.2
python-dotenv==1.0.1
